supervisor/requirements.txt
fastapi
uvicorn
a2a-sdk
langgraph
langchain
langchain_openai
langchain-community
httpx
python-dotenv